backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "artstyle-backend"
version = "0.1.0"
description = "Backend stack for the Art Style Classifier project."
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
  "aio-pika==9.5.5",
  "alembic==1.14.1",
  "asyncpg==0.30.0",
  "boto3==1.37.38",
  "fastapi==0.115.12",
  "greenlet==3.2.1",
  "httpx==0.28.1",
  "mlflow==2.22.0",
  "orjson==3.10.18",
  "pydantic==2.11.4",
  "pydantic-settings==2.9.1",
  "python-multipart==0.0.20",
  "psycopg[binary]==3.2.7",
  "psycopg2-binary==2.9.10",
  "sqlalchemy==2.0.40",
  "uvicorn[standard]==0.34.2",

  # ML model dependencies
  "transformers",
  "Pillow",
  "numpy",
  "scikit-learn",
  "joblib",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-asyncio==0.26.0",
  "ruff==0.11.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/artstyle_backend"]

[tool.hatch.build.targets.wheel.force-include]
"src/artstyle_backend/data/styles.seed.json" = "artstyle_backend/data/styles.seed.json"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
